+++ classes/Helper.php
@@ -34,26 +34,18 @@ class Helper {
         return get_admin_url() . 'admin.php?page=wc-settings&tab=shipping&section=' . self::get_prefix();
     }
 
-    static function print_label_url($shipment_id) {
-        return get_admin_url() . 'admin.php?omniva_global_label=' . $shipment_id;
-    }
-
-    static function print_manifest_labels_url($cart_id = false) {
-        $url = self::generate_manifest_url($cart_id);
-        $url .= '&labels=1';
+    static function generate_outside_action_url($action, $action_value) {
+        $url = admin_url('admin.php?omniva_global_action=' . esc_attr($action) . '&action_value=' . esc_attr($action_value));
         return $url;
     }
 
-    static function generate_manifest_url($cart_id = false) {
-        $url = get_admin_url() . 'admin.php?omniva_global_manifest';
-        if ($cart_id) {
-            $url .= '=' . $cart_id;
-        }
+    static function generate_manifest_page_url($cart_id = false) {
+        $url = admin_url('admin.php?page=omniva_global_manifest');
         return $url;
     }
 
-    static function additional_services() {
-        return [
+    static function additional_services($get_service = false) {
+        $services = array(
             'cod' => __('C.O.D.', 'omniva_global'),
             'return' => __('Return', 'omniva_global'),
             'ukd' => __('U.K.D', 'omniva_global'),
@@ -61,7 +53,9 @@ class Helper {
             'insurance' => __('Insurance', 'omniva_global'),
             'carry_service' => __('Carry service', 'omniva_global'),
             'fragile' => __('Fragile', 'omniva_global')
-        ];
+        );
+
+        return ($get_service && isset($services[$get_service])) ? $services[$get_service] : $services;
     }
 
     static function omniva_get_categories() {
@@ -116,6 +110,16 @@ class Helper {
         }
 
         return $children;
+    }
+
+    static function clear_postcode($postcode, $country) {
+        return $postcode; //This function is not needed yet
+        
+        if ($country == 'LV') {
+            //It is not clear if it will be needed
+        }
+
+        return preg_replace('/[^0-9]/', '', $postcode);
     }
 
 }